requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
itsdangerous==2.2.0
mysqlclient==2.2.7
pillow==11.1.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2024.2
openai>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=14.0.0
azure-storage-blob>=12.17.0 